% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{BatOccu}
\alias{BatOccu}
\title{Occupancy data of 17 bat species.}
\format{A data frame with 49 rows and 51 variables:
\describe{
  \item{MyyuX}{Detection (1) or non-detection (0), of \emph{Myotis yumanensis}:
   on day X}
  \item{MycaX}{Detection (1) or non-detection (0), of \emph{Myotis californicus}:
   on day X}
  \item{MyciX}{Detection (1) or non-detection (0), of \emph{Myotis ciliolabrum}:
   on day X}
  \item{MyvoX}{Detection (1) or non-detection (0), of \emph{Myotis volans}:
   on day X}
  \item{MyluX}{Detection (1) or non-detection (0), of \emph{Myotis lucifugus}:
   on day X}
  \item{PaheX}{Detection (1) or non-detection (0), of \emph{Parastrellus hesperus}:
   on day X}
  \item{LablX}{Detection (1) or non-detection (0), of \emph{Lasiurus blossevillii}:
   on day X}
  \item{MyevX}{Detection (1) or non-detection (0), of \emph{Myotis evotis}:
   on day X}
  \item{AnpaX}{Detection (1) or non-detection (0), of \emph{Antrozous pallidus}:
   on day X}
  \item{EpfuX}{Detection (1) or non-detection (0), of \emph{Eptesicus fuscus}:
   on day X}
  \item{LanoX}{Detection (1) or non-detection (0), of \emph{Lasionycteris noctivagans}:
   on day X}
  \item{MythX}{Detection (1) or non-detection (0), of \emph{Myotis thysanodes}:
   on day X}
  \item{TabrX}{Detection (1) or non-detection (0), of \emph{Tadarida brasiliensis}:
   on day X}
  \item{LaciX}{Detection (1) or non-detection (0), of \emph{Lasiurus cinereus}:
   on day X}
  \item{CotoX}{Detection (1) or non-detection (0), of \emph{Corynorhinus townsendii}:
   on day X}
  \item{EumaX}{Detection (1) or non-detection (0), of \emph{Euderma maculatum}:
   on day X}
  \item{EupeX}{Detection (1) or non-detection (0), of \emph{Eumops perotis}:
   on day X}
}}
\usage{
BatOccu
}
\description{
A dataset containing the detections and non detections of 17 species in the
Plumas County (Columns) for 3 consecutive days in 49 different sites (Rows).
}
\keyword{datasets}

