% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{Dailycov}
\alias{Dailycov}
\title{Nine covariables measured at 49 sampling points in Plumas National Forest, Ca}
\format{A list with 9 elements, each of them with a data frame}
\usage{
Dailycov
}
\description{
A list containing the measurements of nine variables correspondant to 49
different sampling sites (Rows) in Plumas National Forest where the detection
histories of the BatOccu dataset where taken.
}
\details{
\describe{
  \item{Julian}{Ordinal day}
  \item{Maxhum}{Maximum humidity measured on that day}
  \item{Maxtemp}{Maximum temperature meassured on that day}
  \item{Meanhum}{Average humidity meassured on that day}
  \item{Meantemp}{Average temperature meassured on that day}
  \item{Mimhum}{Minimum humidity measured on that day}
  \item{Mintemp}{Minimum temperature meassured on that day}
  \item{sdhum}{Standard deviation of the humidity meassured on that day}
  \item{sdtemp}{Standard devuatuib of the temperature meassured on that day}
}
}
\seealso{
\code{\link[DiversityOccupancy]{BatOccu}}

\code{\link[DiversityOccupancy]{plumas.stack}}

\code{\link[DiversityOccupancy]{sampling.cov}}
}
\keyword{datasets}

