% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotfunctions.R
\name{plot.diversityoccupancy}
\alias{plot.diversityoccupancy}
\title{plots a histogram of diversity of a predicted diversityoccupancy object}
\usage{
\method{plot}{diversityoccupancy}(x)
}
\description{
plots a histogram of diversity of a predicted diversityoccupancy object
}
\examples{
data("BatOccu")
data("Dailycov")
data("sampling.cov")
x <-diversityoccu(pres = BatOccu, sitecov = sampling.cov, obscov = Dailycov,
spp = 17, form = ~ Julian + Meanhum + Meantemp + sdhum + sdtemp ~
Burn.intensity.soil + I(Burn.intensity.soil^2) + Burn.intensity.Canopy +
I(Burn.intensity.Canopy^2) + Burn.intensity.basal +
I(Burn.intensity.basal^2))

plot (x)

}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>
}
\seealso{
\code{\link[DiversityOccupancy]{diversityoccu}}
}

