% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{sampling.cov}
\alias{sampling.cov}
\title{Eight covariables measured at 49 survey stations in the Plumas National Forest}
\format{A data frame with 49 rows and 8 variables:}
\usage{
sampling.cov
}
\description{
A dataframe containing the measurements of 8 variables (columns) correspondant
to 49 different sampling sites (Rows) in the Plumas National Forest where the
detection histories of the BatOccu where taken.
\describe{
  \item{Distance.to.water}{Distance to the closest water source measured in
  meters}
  \item{Distance.to.road}{Distance to the closest road measured in meters}
  \item{Existing.vegetation}{Numeric code to vegetation type}
  \item{Fire.Interval}{Average interval between wildfires, measured in years}
  \item{Altitude}{Metres above sea level of the sampled point}
  \item{Burn.intensity.soil}{This data set reflects fire intensity as
  measured by soil scorched in a scale from 0 to 5}
  \item{Burn.intensity.Canopy}{This data set reflects fire intensity as
  measured by canopy scorched in a scale from 0 to 6}
  \item{Burn.intensity.basal}{This data set reflects fire intensity as
  measured by tree tunks scorched in a scale from 0 to 8}
}
}
\seealso{
\code{\link[DiversityOccupancy]{Dailycov}}

\code{\link[DiversityOccupancy]{BatOccu}}
}
\keyword{datasets}

