% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaryfunctions.R
\name{summary.diversityoccupancy}
\alias{summary.diversityoccupancy}
\title{summarizes the diversity and abundances predicted from a diversityoccupancy
object}
\usage{
\method{summary}{diversityoccupancy}(x)
}
\arguments{
\item{x}{a diversityoccupancy object which is a result from the diversityoccu
function}
}
\description{
summarizes the diversity and abundances predicted from a diversityoccupancy
object
}
\examples{
data("BatOccu")
data("Dailycov")
data("sampling.cov")
x <-diversityoccu(pres = BatOccu, sitecov = sampling.cov, obscov = Dailycov,
spp = 17, form = ~ Julian + Meanhum + Meantemp + sdhum + sdtemp ~
Burn.intensity.soil + I(Burn.intensity.soil^2) + Burn.intensity.Canopy +
I(Burn.intensity.Canopy^2) + Burn.intensity.basal +
I(Burn.intensity.basal^2))

summary(x)

}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>
}
\seealso{
\code{\link[DiversityOccupancy]{diversityoccu}}
}

