% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiversityOccu.R
\name{response.plot}
\alias{response.plot}
\title{plot the response of the calculated alpha diversity to the change of a
particular variable}
\usage{
response.plot(model, variable)
}
\description{
This function takes a model.diversity object and one of the variables used to
predict the
}
\examples{
data("BatOccu")
data("Dailycov")
data("sampling.cov")
x <-diversityoccu(pres = BatOccu, sitecov = sampling.cov, obscov = Dailycov,
spp = 17, form = ~ Julian + Meanhum + Meantemp + sdhum + sdtemp ~
Burn.intensity.soil + I(Burn.intensity.soil^2) + Burn.intensity.Canopy +
I(Burn.intensity.Canopy^2) + Burn.intensity.basal +
I(Burn.intensity.basal^2))
y <- model.diversity(x, method = "g")
response.plot(y, Burn.intensity.soil)
response.plot(y, Existing.vegetation)
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>
}

